detail improvement
